# Improved performance of sum aggregation via aligned loads (-10%)