--- a/src/types/simd/mod.rs
+++ b/src/types/simd/mod.rs
@@ -10,7 +10,10 @@ pub trait FromMaskChunk<T> {
 }
 
 /// A struct that lends itself well to be compiled leveraging SIMD
-pub trait NativeSimd: Default + Copy {
+/// # Safety
+/// The `NativeType` and the `NativeSimd` must have possible a matching alignment.
+/// e.g. slicing `&[NativeType]` by `align_of<NativeSimd>()` must be properly aligned/safe.
+pub unsafe trait NativeSimd: Default + Copy {
     /// Number of lanes
     const LANES: usize;
     /// The [`NativeType`] of this struct. E.g. `f32` for a `NativeSimd = f32x16`.
@@ -32,6 +35,8 @@ pub trait NativeSimd: Default + Copy {
     /// Items from `v` at positions larger than the number of lanes are ignored;
     /// remaining items are populated with `remaining`.
     fn from_incomplete_chunk(v: &[Self::Native], remaining: Self::Native) -> Self;
+
+    fn align(values: &[Self::Native]) -> (&[Self::Native], &[Self], &[Self::Native]);
 }
 
 /// Trait implemented by some [`NativeType`] that have a SIMD representation.