--- a/src/types/simd/native.rs
+++ b/src/types/simd/native.rs
@@ -9,7 +9,7 @@ macro_rules! simd {
         #[derive(Copy, Clone)]
         pub struct $name(pub [$type; $lanes]);
 
-        impl NativeSimd for $name {
+        unsafe impl NativeSimd for $name {
             const LANES: usize = $lanes;
             type Native = $type;
             type Chunk = $mask;
@@ -35,6 +35,11 @@ macro_rules! simd {
                 let mut a = [remaining; $lanes];
                 a.iter_mut().zip(v.iter()).for_each(|(a, b)| *a = *b);
                 Self(a)
+            }
+
+            #[inline]
+            fn align(values: &[Self::Native]) -> (&[Self::Native], &[Self], &[Self::Native]) {
+                unsafe { values.align_to::<Self>() }
             }
         }