This adds aligned load instruction to the SIMD aggregation of arrays without null values. This implementation works on any alignment, and does not require an aligned allocator, and also works on sliced arrays.

Benchmark is a bit mixed on small data sizes. It seems to improving with more data, I want to run one on more data later.

```
Gnuplot not found, using plotters backend
sum 2^10 f32            time:   [101.38 ns 101.41 ns 101.45 ns]                         
                        change: [+1.5061% +1.5401% +1.5774%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 9 outliers among 100 measurements (9.00%)
  2 (2.00%) high mild
  7 (7.00%) high severe

sum 2^12 f32            time:   [284.72 ns 285.05 ns 285.44 ns]                         
                        change: [+0.4929% +0.6963% +0.9300%] (p = 0.00 < 0.05)
                        Change within noise threshold.
Found 7 outliers among 100 measurements (7.00%)
  4 (4.00%) high mild
  3 (3.00%) high severe

sum 2^14 f32            time:   [1.0246 us 1.0246 us 1.0247 us]                          
                        change: [-4.1159% -3.6418% -3.1975%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 7 outliers among 100 measurements (7.00%)
  4 (4.00%) high mild
  3 (3.00%) high severe

sum 2^16 f32            time:   [4.0764 us 4.0773 us 4.0783 us]                          
                        change: [+1.4638% +1.6504% +1.7873%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 9 outliers among 100 measurements (9.00%)
  3 (3.00%) high mild
  6 (6.00%) high severe

sum 2^18 f32            time:   [19.596 us 19.600 us 19.606 us]                          
                        change: [+0.2075% +0.2434% +0.2801%] (p = 0.00 < 0.05)
                        Change within noise threshold.
Found 9 outliers among 100 measurements (9.00%)
  2 (2.00%) high mild
  7 (7.00%) high severe

sum 2^20 f32            time:   [67.890 us 67.909 us 67.931 us]                         
                        change: [-6.5180% -6.4803% -6.4365%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 12 outliers among 100 measurements (12.00%)
  11 (11.00%) high mild
  1 (1.00%) high severe


```